Stub out basic search structure
