+++ requirements.txt
@@ -25,3 +25,4 @@ google-api-python-client==1.5.4
 sqlalchemy==1.1.3
 datefinder==0.6.0
 raven
+pysolr==3.6.0
