+++ search/search.py
@@ -1,0 +1,4 @@
+import pysolr
+
+class Searcher(pysolr.Solr):
+    pass
