+++ sfm_pc/settings.py
@@ -67,6 +67,7 @@ INSTALLED_APPS = (
     'geosite',
     'emplacement',
     'violation',
+    'search',
 )
 
 INSTALLED_APPS += EXTRA_APPS
